backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
pydantic==2.5.1
python-dateutil==2.8.2
python-multipart==0.0.6
langchain>=0.1.0
langchain-groq>=0.1.2
python-dotenv==1.0.0
